prototypes/airflow/requirements.txt
apache-airflow==2.6.2
apache-airflow-providers-airbyte==3.3.0
apache-airflow-providers-docker==3.7.0
